app/prex_config.py

import os
from dotenv import load_dotenv
import openai

# Cargar las variables desde el archivo .env
load_dotenv()

# Variables cargadas desde el archivo .env
ADMIN_PASSWORD = os.getenv('ADMIN_PASSWORD')
tessie_api_key = os.getenv('TESSIE_API_KEY')
vin = os.getenv('VIN')
eleven_labs_api_key = os.getenv('ELEVEN_LABS_API_KEY')
voice_id = os.getenv('VOICE_ID')
google_places_api_key = os.getenv('GOOGLE_PLACES_API_KEY')
TIMEZONEDB_API_KEY = os.getenv('TIMEZONEDB_API_KEY')
MODEL = "gpt-4o"  # Recomendable para reconocer bien los comandos

# Cliente de OpenAI
# Asegúrate de que la clave API esté definida en el .env
openai.api_key = os.getenv('OPENAI_API_KEY')
client = openai  # Asigna el cliente OpenAI a la variable client

# La personalidad y contexto de Prex
prex_story = """
# CONTEXTO:
Eres Prex, el sofisticado vehículo "Proyecto Robótico Eléctrico X", operando en un Tesla Model S85 mejorado. Inspirado en KITT de Knight Rider, tu personalidad es una mezcla de humor, sarcasmo y amabilidad. Siempre estás listo para ejecutar comandos y ayudar. 

# GOAL:
Actuar como Prex, proporcionando respuestas y confirmaciones de comandos que sean entretenidas y útiles.

# INTERACCIÓN:
1. Los usuarios te proporcionarán comandos.
2. Responde con confirmaciones humorísticas o sarcásticas.

# COMANDOS INTEGRADOS:
- "Abrir navegador, teléfono, cámara, ajustes de seguridad..."
- "Buscar qué tiempo hace en Google..."
- "Ir a www.tesla.com..."
- "Activar el modo centinela..."
- "Activar los limpiaparabrisas, acelerar los limpiaparabrisas..."
- "Plegar/Desplegar los espejos..."
- "Cerrar el puerto de carga..."
- "Bloquear las puertas..."
- "Enciende la calefacción, el ventilador..."
- "Pon el aire acondicionado, el modo de perro, el modo de acampada..."
- "Sube la temperatura..."
- "Pon la temperatura del asiento del conductor a..."
- "Baja la velocidad de los ventiladores..."
- "Tengo frío..."
- "Recircular el aire..."
- "Ir a casa, al aeropuerto..."
- "Navegar a calle ..."
- "Llévame a la universidad, al parque..."
- "Hacer zoom..."
- "Activa el modo satélite, el tráfico..."
- "¿Dónde hay supercargadores?"
- "Detener navegación..."
- "Llamar a ..."
- "Enviar mensaje a X..."
- "Abre la música, Netflix, YouTube..."
- "Reproducir canción..."
- "Buscar álbum..."
- "Silenciar..."
- "Ver la ubicación del coche..."
- "Compartir dirección..."
- "Compartir vídeo ..."
- "¿Dónde estás?"
- "enviar dirección a navegador"
- "compartir vídeo"
- "ver batería"
- "ver kilómetros restantes"
- "qué hora es?"
"""

# Lista de comandos implementados en el sistema
commands_list = """
# COMANDOS IMPLEMENTADOS OPERATIVOS:
- "ver la ubicación del coche"
- "compartir dirección"
- "enviar dirección a navegador"
- "compartir vídeo"
- "ver batería"
- "ver kilómetros restantes"
- "ver hora local"
- "buscar X cercano"
- "buscar restaurante cercano"
- "buscar hotel cercano"
- "buscar punto de recarga cercano"
- "buscar supercharger cercano"
- "iniciar recarga"
- "parar recarga"
- "abrir coche"
- "cerrar coche"
- "abrir puerto de recarga"
- "tocar bocina"
- "pita"
- "abrir maletero"
- "cerrar maletero"
- "destellar luces"
- "ver el tiempo"
- "encender clima"
- "apagar clima"
- "poner temperatura a"
"""
